Service Servers and Clients Nodes
"7.3. Creating and Running Service Servers and Client Nodes
A service can be divided into two types: a Service Server that responds only when there is a
request and a Service Client that can send both requests and respond to requests. Unlike the
topic, the service is a one-time message communication. Therefore, when the request and the
response of the service are completed, the two connected nodes will be disconnected.
These services are often used when requesting a robot to perform a specific action.
Alternatively, it is used for nodes that require specific events to occur under certain conditions.
As service is a single occurrence of communication method, it is a very useful method that can
replace the topic with a small network bandwidth.
In this section, we will create a simple service file and run a service server node and a service
client node."----(ROS_Robot_Programming, author: Dr. Pyo, chapter 7.3)