Action Server and Client Node
7.4. Writing and Running the Action Server and Client Node
In this section, the action server and action client nodes were created for experimenting the
''Action'', which is the third message communication method. Unlike
topics and services, actions are very useful for asynchronous, bidirectional, and more complex
programming where an extended response time is expected, after processing request and
intermediate feedback are needed.