Chapter7.6 roslaunch
The ‘rosrun’ is a command that executes just one node, and ‘roslaunch’ can run more than one
node. Other features of the ‘roslaunch’ command include the ability to modify parameters of the
package, rename the node name, the ROS_ROOT and ROS_PACKAGE_PAPATION_PATH settings,
and change environment variables.
The ‘roslaunch’ uses the ‘*.launch’ file to select executable nodes, which is XML-based and
provides tag-specific options. The execution command is ‘roslaunch [package name] [roslaunch
file]’.


<launch>
<group ns="ns1">
<node pkg="ros_tutorials_topic" type="topic_publisher" name="topic_publisher"/>
<node pkg="ros_tutorials_topic" type="topic_subscriber" name="topic_subscriber"/>
</group>
<group ns="ns2">
<node pkg="ros_tutorials_topic" type="topic_publisher" name="topic_publisher"/>
<node pkg="ros_tutorials_topic" type="topic_subscriber" name="topic_subscriber"/>
</group>
</launch>